refactor(engineio): distinct server and services modules
