+++ engineioxide/Cargo.toml
@@ -61,7 +61,7 @@ hyper = { version = "0.14.25", features = [
 ] }
 
 [features]
-default = ["v4", "hyper-v1"]
+default = ["v4"]
 v4 = []
 v3 = ["memchr", "unicode-segmentation"]
 test-utils = []
